[All] in order to have different packages, we need to take care that we don't violate our dependency inclusions. Therefore this refactors includes and moves some files around to fix those violations. this also solves a lot of recursive dependency issues like OrderComponent <-> PaymentComponent.
