+++ composer.json
@@ -100,7 +100,8 @@
     "friends-of-behat/cross-container-extension": "^1.1",
     "friends-of-behat/service-container-extension": "^1.0",
     "friends-of-behat/symfony-extension": "^1.2",
-    "friends-of-behat/variadic-extension": "^1.1"
+    "friends-of-behat/variadic-extension": "^1.1",
+    "mamuz/php-dependency-analysis": "^1.3"
   },
   "autoload": {
     "psr-4": {
