+++ src/CoreShop/Bundle/CoreBundle/Form/Extension/CustomerTypeExtension.php
@@ -1,0 +1,58 @@
+<?php
+/**
+ * CoreShop.
+ *
+ * This source file is subject to the GNU General Public License version 3 (GPLv3)
+ * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
+ * files that are distributed with this source code.
+ *
+ * @copyright  Copyright (c) 2015-2017 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
+ * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
+ */
+
+namespace CoreShop\Bundle\CoreBundle\Form\Extension;
+
+use CoreShop\Bundle\AddressBundle\Form\Type\SalutationChoiceType;
+use CoreShop\Bundle\CoreBundle\Form\Type\AddressChoiceType;
+use CoreShop\Bundle\CustomerBundle\Form\Type\CustomerType;
+use Symfony\Component\Form\AbstractTypeExtension;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+final class CustomerTypeExtension extends AbstractTypeExtension
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder->add('salutation', SalutationChoiceType::class, [
+                'label' => 'coreshop.form.customer.salutation'
+            ]);
+
+        if ($options['allow_default_address'] && $options['customer']) {
+            $builder->add('defaultAddress', AddressChoiceType::class, [
+                'customer' => $options['customer'],
+                'label' => 'coreshop.form.customer.default_address',
+            ]);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        parent::configureOptions($resolver);
+
+        $resolver->setDefault('allow_default_address', false);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getExtendedType()
+    {
+        return CustomerType::class;
+    }
+}
