+++ src/CoreShop/Component/Inventory/composer.json
@@ -19,7 +19,8 @@
   ],
   "require": {
     "php": "^7.0",
-    "coreshop/resource": "^2.0"
+    "coreshop/resource": "^2.0",
+    "coreshop/rule": "^2.0"
   },
   "config": {
     "bin-dir": "bin"
