+++ src/CoreShop/Component/Order/Transformer/OrderToOrderDocumentTransformerWorkflowApplier.php
@@ -2,9 +2,9 @@
 
 namespace CoreShop\Component\Order\Transformer;
 
-use CoreShop\Bundle\WorkflowBundle\Manager\StateMachineManager;
 use CoreShop\Component\Order\Model\OrderDocumentInterface;
 use CoreShop\Component\Order\Model\OrderInterface;
+use CoreShop\Bundle\WorkflowBundle\Manager\StateMachineManagerInterface;
 
 final class OrderToOrderDocumentTransformerWorkflowApplier implements OrderDocumentTransformerInterface
 {
@@ -14,7 +14,7 @@ final class OrderToOrderDocumentTransformerWorkflowApplier implements OrderDocum
     protected $innerTransformer;
 
     /**
-     * @var StateMachineManager
+     * @var StateMachineManagerInterface
      */
     protected $stateMachineManager;
 
@@ -36,12 +36,12 @@ final class OrderToOrderDocumentTransformerWorkflowApplier implements OrderDocum
     /**
      * OrderToInvoiceTransformerWorkflowApplier constructor.
      * @param OrderDocumentTransformerInterface $innerTransformer
-     * @param StateMachineManager $stateMachineManager
+     * @param StateMachineManagerInterface $stateMachineManager
      * @param string $initialState
      * @param string $workflowName
      * @param string $transition
      */
-    public function __construct(OrderDocumentTransformerInterface $innerTransformer, StateMachineManager $stateMachineManager, string $initialState, string $workflowName, string $transition)
+    public function __construct(OrderDocumentTransformerInterface $innerTransformer, StateMachineManagerInterface $stateMachineManager, string $initialState, string $workflowName, string $transition)
     {
         $this->innerTransformer = $innerTransformer;
         $this->stateMachineManager = $stateMachineManager;
