+++ src/CoreShop/Component/Order/composer.json
@@ -19,17 +19,20 @@
   ],
   "require": {
     "php": "^7.0",
+    "coreshop/address": "^2.0",
     "coreshop/rule": "^2.0",
     "coreshop/registry": "^2.0",
     "coreshop/resource": "^2.0",
+    "coreshop/sequence": "^2.0",
     "coreshop/store": "^2.0",
     "coreshop/currency": "^2.0",
     "coreshop/taxation": "^2.0",
     "coreshop/payment": "^2.0",
-    "coreshop/address": "^2.0",
+    "coreshop/pimcore": "^2.0",
     "coreshop/customer": "^2.0",
     "coreshop/storage-list": "^2.0",
-    "coreshop/locale": "^2.0"
+    "coreshop/locale": "^2.0",
+    "coreshop/workflow-bundle": "^2.0"
   },
   "config": {
     "bin-dir": "bin"
