+++ src/CoreShop/Component/Payment/Model/PayableInterface.php
@@ -10,9 +10,9 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-namespace CoreShop\Bundle\CoreBundle;
+namespace CoreShop\Component\Payment\Model;
 
-final class CoreEvents
+interface PayableInterface
 {
-    const SUPPORTS_PAYMENT_PROVIDER = 'coreshop.payment_provider.supports';
+    public function getId();
 }
